+++ addons/antennas/config.cpp
@@ -14,7 +14,8 @@ class CfgPatches {
             "acre_sys_gsa",
             "acre_ace_interact",
             "A3_Props_F_Enoch_Military_Camps",
-            "A3_Props_F_Enoch_Military_Equipment"
+            "A3_Props_F_Enoch_Military_Equipment",
+            "A3_Props_F_Decade_Objectives"
         };
         // Skip this addon instead of erroring when ACRE, ACE or Contact content is missing
         skipWhenMissingDependencies = 1;